+++ src/renderer/src/views/components/panels/co-output-panel.vue
@@ -53,7 +53,7 @@ import type { Output } from '@/types';
 import CoButton from '@/components/co-button/index.vue';
 import { Delete } from '@/components/co-icon/index';
 import CoSettingsPanel from '@/components/co-settings-panel/index.vue';
-import { injectCoPic } from '@renderer/uses';
+import { injectCoPic } from '@renderer/uses/co-pic';
 import { storage } from '@renderer/utils/storage';
 import { ref, watch } from 'vue';
